+++ PSFramework/changelog.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.9.8.16 : XXXX-XX-XX
+ - New: Added command Register-PSFParameterClassMapping
+ - New: Added command Get-PSFTypeSerializationData
+ - New: Added command Register-PSFTypeSerializationData
+ - New: Added class `[PSFramework.Serialization.SerializationTypeConverter]`, a type serializer that can be used to serialize and deserialize types.
+ - Upd: Messages now also include the file and line they were written in
+
 ## 0.9.7.14 : 2018-01-17
  - New: Added tests to module
  - New: Added parameter class: ComputerParameter (`[PSFComputer]`)